+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 - **New**
     - Add COM server initialization
     - Add COM server initialization tests
+    - Add list methods of COM instance
+    - Add get method of COM instance